[HIR] factored out common code in integration tests.
